requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyMuPDF>=1.23.7
Pillow>=10.2.0
python-magic>=0.4.27
transformers>=4.36.2
torch>=2.1.2
numpy>=1.24.3